change contact

+++ service/modules/domain/src/main/java/dev/fizlrock/waterwalk/domain/values/Distance.java
@@ -3,23 +3,23 @@ package dev.fizlrock.waterwalk.domain.values;
 /** Расстояние в метрах */
 public final class Distance {
 
-  private final Long distance;
+  private final int distance;
 
-  public Distance(Long distanceInMeters) {
+  public Distance(int distanceInMeters) {
     if (distanceInMeters <= 0)
       throw new IllegalArgumentException("Дистанция должна быть больше нуля");
     this.distance = distanceInMeters;
   }
 
-  public static Distance ofMeters(long distanceInMeters) {
+  public static Distance ofMeters(int distanceInMeters) {
     return new Distance(distanceInMeters);
   }
 
-  public static Distance ofKMeters(long distanceInKMeters) {
+  public static Distance ofKMeters(int distanceInKMeters) {
     return new Distance(distanceInKMeters * 1000);
   }
 
-  public Long inMeters() {
+  public int inMeters() {
     return distance;
   }
